remove API dependecy from drag & drop composable
